requirements.txt
fastapi
uvicorn
supabase
pdfplumber
PyMuPDF
markdownify
requests
python-docx
pydantic
python-dotenv
camelot-py   
pytesseract
PyYAML
numpy
langchain
langchain-core
langchain-community
langchain-text-splitters
unstructured[md]
nltk
python-dotenv

openai
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
